requirements_dev.txt
bumpversion==0.5.3
coverage==4.5.1
flake8==3.8.3
pip==18.1
python-coveralls==2.9.1
sphinx==1.8.1
sphinx_rtd_theme==0.4.3
tox==3.5.2
twine==1.12.1
watchdog==0.9.0
wheel==0.32.1
django-germanium>=2.2.3
